requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
tabula-py
openpyxl